vision/requirements.txt
opencv-python
pytesseract
pillow
numpy
fastapi
uvicorn[standard]
